Remove the "all signals" logic
The filter combinator use to use a constant combinator internally that
had "all signals" set to it. This became cumbersome with 2.0 where not
only a lot of signals exist but also each signal is multiplied by the
number of quality versions. This led to a very noticable delay when
computing this list and assigning it to a constant combinator (which can
do 10,000 signals but it is slow).